+++ exampleCourse/questions/brokenGeneration/server.py
@@ -1,0 +1,2 @@
+def generate(data):
+    raise Exception('deliberately broken generate function')