+++ ChangeLog.md
@@ -27,6 +27,8 @@
 
   * Change external grading to receive results from an SQS queue instead of a webhook (Nathan Walters).
 
+  * Change Exam question generation to first-access time (Matt West).
+
   * Fix load-reporting close during unit tests (Matt West).
 
   * Fix PL / scheduler linking stored procedure to allow linked exams and fix bugs (Dave Mussulman).